requirements.txt
alembic==1.2.1
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.2.0
blinker==1.4
Click==7.0
colorama==0.4.1
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
psycopg2==2.8.3
py==1.8.0
pylint==2.3.1
pyparsing==2.4.2
pytest==5.2.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.6
typed-ast==1.4.0
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
zipp==0.6.0
